@@ -11,40 +11,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.3",
-    "@angular/cdk": "^7.0.3",
-    "@angular/common": "^7.0.3",
-    "@angular/compiler": "^7.0.3",
-    "@angular/core": "^7.0.3",
-    "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "^7.0.3",
-    "@angular/http": "^7.0.3",
-    "@angular/material": "^7.0.3",
-    "@angular/platform-browser": "^7.0.3",
-    "@angular/platform-browser-dynamic": "^7.0.3",
-    "@angular/router": "^7.0.3",
-    "@ngrx/effects": "^7.0.0-beta.0",
-    "@ngrx/entity": "^7.0.0-beta.0",
-    "@ngrx/store": "^7.0.0-beta.0",
-    "@ngrx/store-devtools": "^7.0.0-beta.0",
-    "core-js": "^2.5.4",
-    "dayjs": "^1.7.7",
-    "firebase": "^5.5.3",
+    "@angular/animations": "^7.1.4",
+    "@angular/cdk": "^7.2.0",
+    "@angular/common": "^7.1.4",
+    "@angular/compiler": "^7.1.4",
+    "@angular/core": "^7.1.4",
+    "@angular/flex-layout": "^7.0.0-beta.22",
+    "@angular/forms": "^7.1.4",
+    "@angular/http": "^7.1.4",
+    "@angular/material": "^7.2.0",
+    "@angular/platform-browser": "^7.1.4",
+    "@angular/platform-browser-dynamic": "^7.1.4",
+    "@angular/router": "^7.1.4",
+    "@ngrx/effects": "^7.0.0",
+    "@ngrx/entity": "^7.0.0",
+    "@ngrx/store": "^7.0.0",
+    "@ngrx/store-devtools": "^7.0.0",
+    "core-js": "^2.6.1",
+    "dayjs": "^1.7.8",
+    "firebase": "^5.7.2",
     "hammerjs": "^2.0.8",
-    "highlight.js": "^9.12.0",
+    "highlight.js": "^9.13.1",
     "lodash": "^4.17.11",
-    "marked": "^0.5.0",
+    "marked": "^0.5.2",
     "rxjs": "~6.3.3",
-    "rxjs-compat": "^6.3.2",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.5",
-    "@angular/compiler-cli": "^7.0.3",
-    "@angular/language-service": "^7.0.3",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "^0.11.4",
+    "@angular/cli": "^7.0.7",
+    "@angular/compiler-cli": "^7.1.4",
+    "@angular/language-service": "^7.1.4",
+    "@types/jasmine": "^2.8.14",
+    "@types/jasminewd2": "^2.0.6",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.3.0",
     "jasmine-core": "~2.99.1",
@@ -54,7 +54,7 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "protractor": "^5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
     "typescript": "~3.1.6"
