@@ -5,12 +5,14 @@ import { select, Store } from '@ngrx/store';
 import { of as observableOf } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { Subject } from 'rxjs/Subject';
-import { takeUntil } from 'rxjs/operators';
-import { find, get, pull } from 'lodash';
+import { takeUntil, skip } from 'rxjs/operators';
+import { find, get, pull, isEmpty, every } from 'lodash';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 
 import { Blog } from '../../model/blog';
 import * as fromBlog from '../../reducer';
 import * as BlogActions from '../../actions/blog.actions';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-edit-blog',
@@ -33,10 +35,12 @@ export class EditBlogComponent implements OnInit, OnDestroy {
     private fb: FormBuilder,
     private route: ActivatedRoute,
     private router: Router,
-    private store: Store<fromBlog.State>
+    private store: Store<fromBlog.State>,
+    public dialog: MatDialog,
   ) {}
 
   ngOnInit(): void {
+    this.handleError();
     this.route.paramMap
       .pipe(
         switchMap((params: ParamMap) => {
@@ -102,6 +106,9 @@ export class EditBlogComponent implements OnInit, OnDestroy {
 
   save(isDraft: boolean): void {
     const formValue = this.form.getRawValue();
+    if (every(formValue, isEmpty)) {
+      return;
+    }
     const blog = {
       isDraft,
       title: formValue.title,
@@ -114,16 +121,6 @@ export class EditBlogComponent implements OnInit, OnDestroy {
     } else {
       this.store.dispatch(new BlogActions.AddBlog({ ...blog, createdDate: new Date().valueOf() }));
     }
-    this.store
-      .pipe(
-        select(fromBlog.getErrorMessage),
-        takeUntil(this.destroy$)
-      )
-      .subscribe((errorMessage) => {
-        if (errorMessage === 'Success') {
-          this.router.navigate(['/blog']);
-        }
-      });
   }
 
   back(): void {
@@ -132,6 +129,44 @@ export class EditBlogComponent implements OnInit, OnDestroy {
 
   togglePreview(): void {
     this.showPreview = !this.showPreview;
+  }
+
+  delete(): void {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '400px',
+      data: {
+        title: '警告',
+        message: '确认删除这篇文章吗？',
+        confirm: {
+          text: '删除',
+          color: 'warn',
+        },
+        cancel: {
+          text: '取消',
+          color: 'primary',
+        },
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (!!this.blogId && confirmed) {
+        this.store.dispatch(new BlogActions.DeleteBlog({ blog: this.blog }));
+      }
+    });
+  }
+
+  handleError(): void {
+    this.store
+    .pipe(
+      select(fromBlog.getErrorMessage),
+      skip(1),
+      takeUntil(this.destroy$)
+    )
+    .subscribe((errorMessage) => {
+      if (errorMessage === 'Success') {
+        this.router.navigate(['/blog']);
+      }
+    });
   }
 
   ngOnDestroy(): void {
