@@ -2,8 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 const appRoutes: Routes = [
-  { path: '', redirectTo: '/blog', pathMatch: 'full' },
-  { path: '**', redirectTo: 'blog' }
+  { path: '', redirectTo: 'blog', pathMatch: 'full' },
 ];
 
 @NgModule({
